requirements/local.txt
pytest
pytest-cov
mock
pytest-django
pytest-mock
flake8
